REDCap_HealthPro_Sync/requirements.txt
certifi==2019.11.28
chardet==3.0.4
future==0.18.2
gspread==3.1.0
httplib2==0.16.0
idna==2.8
numpy==1.18.1
oauth2client==4.1.3
pandas==0.25.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyCap==0.0.0
python-dateutil==2.8.1
pytz==2019.3
redcaphelper==0.2.4.dev0
requests==2.22.0
rsa==4.0
semantic-version==2.3.1
six==1.14.0
urllib3==1.25.7
virtualenv==16.7.9
